+++ backend/test/src/Flow/HighTongueTest.hs
@@ -24,7 +24,7 @@ testSupportedDSLVersion = assertRight $ decodeHighTongue process
   where
     -- TODO: use a version from HighTongue.supportedVersions in "process"
     process = unsafeProcess [r|
-          dsl-version: "0.1.0"
+          dsl-version: "0.2.0"
           stages:
             - sign:
                 actions: []
@@ -66,7 +66,7 @@ testDuplicateStageNameErrors = assertLeft $ decodeHighTongue process
   where
     -- TODO: use a version from HighTongue.supportedVersions in "process"
     process = unsafeProcess [r|
-          dsl-version: "0.1.0"
+          dsl-version: "0.2.0"
           stages:
             - dup:
                 actions: []